[OpenCL] Adds SYCL version of ApplyAdagrad
Need a slightly different version of ApplyAdagrad to the CPU version, as
the scalar lr should be on the device, but we cannot easily multiply a
tensor by a scalar which does not exist in host memory. Instead we
broadcast the scalar into a suitably sized tensor and multiply these
together to get the same result.