[OpenCL] Changes expected NaN behaviour in test
The new SYCL kernels provide the same behaviour as the CUDA and cuDNN
kernels when an input tensor only contains NaN and the test needs to
reflect this.

As NaN cannot be compared to any other float value, it makes little
sense to decide which of the NaNs is the maximum, and so which NaN
should have the error propagated to it.