[OpenCL] Provides atomic free MaxPool gradients (#110)
* [OpenCL] Provides atomic free MaxPool3DGrad

Atomic support in SYCL is not designed in a way that plays nicely with
Tensorflow and Eigen. Here we provide a new implementation for
MaxPool3DGrad which does not rely on atomics, and so avoids any such
problems.

* [OpenCL] Provides atomic free MaxPoolGrad

Atomic support in SYCL is not designed in a way that plays nicely with
Tensorflow and Eigen. Here we provide a new implementation for
MaxPoolGrad which does not rely on atomics, and so avoids any such
problems.

* [OpenCL] Changes expected NaN behaviour in test

The new SYCL kernels provide the same behaviour as the CUDA and cuDNN
kernels when an input tensor only contains NaN and the test needs to
reflect this.

As NaN cannot be compared to any other float value, it makes little
sense to decide which of the NaNs is the maximum, and so which NaN
should have the error propagated to it.

* [OpenCL] Removes unneeded SYCL atomic functions

* [OpenCL] Tidies SYCL MaxPoolGrad kernels

Some tidying up and also adds a local accumulator value which will be
written to memory at the end of the kernel, to decrease the number og
memory writes in the kernel.