[OpenCL] Fixes expected cost models with devices
The aim of this test is to ensure that there is at least one cost model
(that of the CPU), as if there are no cost models then the test should
fail and there is no point in trying to run the other tests on the cost
model.

The number of cost models provided by a session should matches the
number of devices used by that session. When additional devices are
present there will be more than one cost model, so we should not be
checking for equality, but rather that the number of cost models is at
least one.