[OpenCL] Fix allocator destruction race condition
A Tensor's allocator must outlive it, however there is no easy way to
determine whether an Allocator has any Tensors still alive, and so we
cannot know when it is safe to destroy an allocator. The CPU allocator
gets round this by being deleted, so we adopt this convention here.