Further improvements to the crosscompile toolchain
Now there is a templated bash script that sets the target, gcc-toolchain
and sysroot for compute (the c-like compute++). This means that we no
longer need the version flag, as clang will basically work it out for
us. Similarly, all the builtin directories and include lines have been
removed from the CROSSTOOL as they're not needed any more!

This successfully compiles the host and device objects, but fails to link
because it is not aware that two libComputeCpp.so files are needed - one
for host, and one for device.