[SYCL] Bazel header dependency hotfix
Add the canonical header flag to gcc when compiling non-SYCL C++. This
flag is required to ensure that gcc will report the dependencies using
the directories provided by bazel, whereas by default gcc will use
a relative or absolute path depending on which is shortest. In general
this does not cause a problem, as the bazel cache has a fairly long
name and so the relative path is almost always shorter than the absolute
path, however this might not be the case when you change the location of
the bazel cache.

Regession introduced: bfaac96
Fixes: #215
See also: https://github.com/bazelbuild/bazel/issues/1642