[SYCL] Add SYCL bitcode configure option
Add a configure time option to specify the SYCL target bitcode to pass
to ComputeCpp. By default this is spir64 for AMD devices, but it can be
changed to match the actual device to compiler for.

Note: There is no check to ensure that the target is supported by
ComputeCpp, nor is there a check to ensure that the traget matches the
device.